Try to import librecaptcha.__init__ instead
When this repository is cloned, there is initially an empty "librecaptcha"
directory. Python will detect this as a namespace package and successfully
import it, but we want a ModuleNotFoundError to be thrown so the submodule
can be cloned.

However, once "librecaptcha/" is inserted into sys.path and downloaded,
"librecaptcha" will still be cached as a namespace package, so we must call
importlib.reload() to load it from the correct location.